+++ api/proto/buf.gen.yaml
@@ -3,7 +3,7 @@ version: v1
 managed:
   enabled: true
   go_package_prefix:
-    default: github.com/traPtitech/Emoine_R/pkg/bufgen
+    default: github.com/traPtitech/Emoine_R/pkg/pbgen
 
 plugins:
   - name: go
@@ -11,14 +11,14 @@ plugins:
       - go
       - run
       - google.golang.org/protobuf/cmd/protoc-gen-go@latest
-    out: ../../pkg/bufgen
+    out: ../../pkg/pbgen
     opt: paths=source_relative
   - name: connect-go
     path:
       - go
       - run
       - github.com/bufbuild/connect-go/cmd/protoc-gen-connect-go@latest
-    out: ../../pkg/bufgen
+    out: ../../pkg/pbgen
     opt: paths=source_relative
   - name: doc
     path:
